Expressions: Reduce constants and prune branches in logic ops
Previously, constant expressions were fully compiled (to know that the
compiler is doing the right thing), but now on this stage we can reduce
those constants and avoid unneeded instructions.

Furthermore, this has enabled an optimization on `||` and `&&`
operators, pruning the other half if one half is constant and already
defines the result of the expression.