bug(variables): Detect uses of variables in init expressions of other variables
The compiler couldn't recognize the use of one variable in the
initialization of another as its first use, that is:
```c
int a;
int b = a = 0;
```
this could not be recognized, and `a` was flagged as unused, which later
results in a panic for not having `a` when trying to assign indices to
variables.

Now this case is taken into account as well by counting references for
existing init expressions in variable declarations.