refactor: move all current optimizations to execute in a single place
This was motivated by trying to avoid redundancy in the compiler with
`cset`s, because constant propagation removes all the "special" tags
from the expressions as constant propagation should run **before**
codegen and not during. This was not the case so I separated all the
transformations to a `constant_folding` module that runs **once** on
each block, doing all of the permutations it can in the way, and the
HLIR doesn't re-visit that function once it has passed through it so the
code generator can freely add tags to the expressions without those tags
being removed.