ir/fold: Add better propagation mechanism
Now each of the propagation components (Value, Statement) return
a `PropagationResult` which encodes the new value and a boolean
containing whether or not a modification was made at some point.

This allows to know whether a certain fold succeeded or not while
folding a block and therefore be able to retry previously failed
folds in the same statement if another fold succeeded in said statement.