bug(fold): Correctly inline phi nodes when selecting branches
The inliner did the pass that finds phi nodes and sets them to the
value from the selected branch after the block was renamed to its
parent, which creates situations that make the phi node have various
values attached to the same block, which invalidates them. Now that
specific pass has been set to run before any changes go through the code
when inlining.

Also the pruning of unreached blocks is done separately 1) because it is
a heavyweight operation (may require a lot of renaming passes) and 2)
because there are only two places when a pass like that is relevant:
  - before we start folding, because code given by frontend is very poor
  - after folding all the blocks, where there might have been branch
    changes and the branching graphs might have changed.