flags: Force returned bindings on registers
There might be places where the state of the CPU hasn't been lost yet
and we can use the binding as a flag as well in branches, but this is an
easier fix for the moment. Tracking how CPU flags flow through the
program isn't complicated though, so might be done later.